Move utility functions to string_refinement_util
This reduces the size of string_refinement files and seperate functions
that could be reused outside of string_refinement, for instace for unit testing.
