+++ src/solvers/refinement/string_refinement.h
@@ -93,12 +93,6 @@ exprt concretize_arrays_in_expression(
   std::size_t string_max_length,
   const namespacet &ns);
 
-bool is_char_array_type(const typet &type, const namespacet &ns);
-
-bool has_subtype(
-  const typet &type,
-  const std::function<bool(const typet &)> &pred);
-
 // Declaration required for unit-test:
 union_find_replacet string_identifiers_resolution_from_equations(
   std::vector<equal_exprt> &equations,
