+++ src/solvers/refinement/string_refinement_util.h
@@ -11,6 +11,62 @@
 
 #include "string_constraint.h"
 
+/// For now, any unsigned bitvector type of width smaller or equal to 16 is
+/// considered a character.
+/// \note type that are not characters maybe detected as characters (for
+/// instance unsigned char in C), this will make dec_solve do unnecessary
+/// steps for these, but should not affect correctness.
+/// \param type: a type
+/// \return true if the given type represents characters
+bool is_char_type(const typet &type);
+
+/// Distinguish char array from other types.
+/// For now, any unsigned bitvector type is considered a character.
+/// \param type: a type
+/// \param ns: namespace
+/// \return true if the given type is an array of characters
+bool is_char_array_type(const typet &type, const namespacet &ns);
+
+/// For now, any unsigned bitvector type is considered a character.
+/// \param type: a type
+/// \return true if the given type represents a pointer to characters
+bool is_char_pointer_type(const typet &type);
+
+/// \param type: a type
+/// \param pred: a predicate
+/// \return true if one of the subtype of `type` satisfies predicate `pred`.
+///         The meaning of "subtype" is in the algebraic datatype sense:
+///         for example, the subtypes of a struct are the types of its
+///         components, the subtype of a pointer is the type it points to,
+///         etc...
+///         For instance in the type `t` defined by
+///         `{ int a; char[] b; double * c; { bool d} e}`, `int`, `char`,
+///         `double` and `bool` are subtypes of `t`.
+bool has_subtype(
+  const typet &type,
+  const std::function<bool(const typet &)> &pred);
+
+/// \param type: a type
+/// \return true if a subtype of `type` is an pointer of characters.
+///         The meaning of "subtype" is in the algebraic datatype sense:
+///         for example, the subtypes of a struct are the types of its
+///         components, the subtype of a pointer is the type it points to,
+///         etc...
+bool has_char_pointer_subtype(const typet &type);
+
+/// \param type: a type
+/// \return true if a subtype of `type` is string_typet.
+///         The meaning of "subtype" is in the algebraic datatype sense:
+///         for example, the subtypes of a struct are the types of its
+///         components, the subtype of a pointer is the type it points to,
+///         etc...
+bool has_string_subtype(const typet &type);
+
+/// \param expr: an expression
+/// \param ns: namespace
+/// \return true if a subexpression of `expr` is an array of characters
+bool has_char_array_subexpr(const exprt &expr, const namespacet &ns);
+
 struct index_set_pairt
 {
   std::map<exprt, std::set<exprt>> cumulative;
