+++ src/util/magic.h
@@ -8,5 +8,6 @@
 #include <cstddef>
 
 const std::size_t MAX_FLATTENED_ARRAY_SIZE=1000;
+const std::size_t STRING_REFINEMENT_MAX_CHAR_WIDTH = 16;
 
 #endif
